change tag control name and icon to search
