+++ demo/advanced-styling/index.html
@@ -62,7 +62,7 @@
         border-color: orange;
       }
 
-      body .meridian-tag-control,
+      body .meridian-search-control,
       body .meridian-floor-control {
         border-color: darkgreen;
         border-radius: 0 10px 0 10px;
@@ -71,9 +71,9 @@
         box-shadow: 0 0 3px 3px orange;
       }
 
-      body .meridian-tag-control:hover,
+      body .meridian-search-control:hover,
       body .meridian-floor-control:hover,
-      body .meridian-tag-control:focus,
+      body .meridian-search-control:focus,
       body .meridian-floor-control:focus {
         outline: none;
         background: mediumturquoise;
