+++ demo/dark-theme/index.html
@@ -39,7 +39,7 @@
       }
 
       .map .meridian-floor-control,
-      .map .meridian-tag-control,
+      .map .meridian-search-control,
       .map .meridian-zoom-button {
         fill: #fff;
         background: #444;
@@ -49,7 +49,7 @@
       }
 
       .map .meridian-floor-control:focus,
-      .map .meridian-tag-control:focus,
+      .map .meridian-search-control:focus,
       .map .meridian-zoom-button:focus {
         box-shadow: 0 0 0 3px #0f0, 0 0 3px #0f0;
       }
@@ -64,13 +64,13 @@
       }
 
       .map .meridian-floor-control:hover,
-      .map .meridian-tag-control:hover,
+      .map .meridian-search-control:hover,
       .map .meridian-zoom-button:hover {
         background: #555;
       }
 
       .map .meridian-floor-control:active,
-      .map .meridian-tag-control:active,
+      .map .meridian-search-control:active,
       .map .meridian-zoom-button:active {
         background: #666;
       }
